Fikser utdatert amplitude og går over til analytics
